removed some files and folders, restructuring to a smoother heirarchy and layout
