--- client/src/Environment/Basic.jsx
@@ -1,21 +1,0 @@
-const Basic = () => {
-    return (
-        <div>
-           <section>
-                <p>Homebrew</p>
-           </section>
-
-           <section>
-                <p>Visual Studio Code</p>
-           </section>
-
-           <section>
-                <p>Bundler</p>
-           </section>
-        </div>
-    )
-}
-
-export default Basic;
-
-
