--- client/src/Environment/Ruby.jsx
@@ -1,7 +1,0 @@
-const Ruby = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Ruby;
