--- client/src/FrameworkList/Rails.jsx
@@ -1,7 +1,0 @@
-const Rails = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Rails;
