--- client/src/FrameworkList/React.jsx
@@ -1,7 +1,0 @@
-const React = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default React;
