--- client/src/FrameworkList/Sinatra.jsx
@@ -1,7 +1,0 @@
-const Sinatra = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Sinatra;
