--- client/src/Frameworks.jsx
@@ -1,7 +1,0 @@
-const Frameworks = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Frameworks;
