--- client/src/LanguageList/Java.jsx
@@ -1,7 +1,0 @@
-const Java = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Java;
