--- client/src/LanguageList/JavaScript..jsx
@@ -1,7 +1,0 @@
-const JavaScript = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default JavaScript;
