--- client/src/LanguageList/Python.jsx
@@ -1,7 +1,0 @@
-const Python = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Python;
