--- client/src/Languages.jsx
@@ -1,12 +1,0 @@
-const Languages = () => {
-    return (
-        <div>
-            <button>Ruby</button>
-            <button>JavaScript</button>
-            <button>Python</button>
-            <button>Java</button>
-        </div>
-    )
-}
-
-export default Languages;
