+++ client/src/ParentComponents/Main/Database.jsx
@@ -1,0 +1,9 @@
+const Database = () => {
+    return (
+        <div>
+            <h1>Database Setup Here</h1>
+        </div>
+    )
+}
+
+export default Database;
