+++ client/src/ParentComponents/Main/Environment.jsx
@@ -1,0 +1,9 @@
+const Environment = () => {
+    return (
+        <div>
+            <h1>Environment Setup Here</h1>
+        </div>
+    )
+}
+
+export default Environment;
