+++ client/src/ParentComponents/Main/Framework.jsx
@@ -1,0 +1,9 @@
+const Framework = () => {
+    return (
+        <div>
+            <h1>Framework Setup Here</h1>
+        </div>
+    )
+}
+
+export default Framework;
