+++ client/src/ParentComponents/Main/Languages.jsx
@@ -1,0 +1,9 @@
+const Languages = () => {
+    return (
+        <div>
+            <h1>Language Setup Here</h1>
+        </div>
+    )
+}
+
+export default Languages;
